LNP-1580 | Refactor Playwright test configuration and structure for improved organization and clarity
