+++ package.json
@@ -3,13 +3,13 @@
   "private": true,
   "description": "Playwright end-to-end tests for prisoner-content-hub-backend",
   "scripts": {
-    "test:e2e": "playwright test",
-    "test:e2e:all": "playwright test --workers=1",
-    "test:e2e:headed": "playwright test --headed",
-    "test:e2e:debug": "playwright test --debug",
-    "test:e2e:ui": "playwright test --ui",
+    "test:e2e": "playwright test -c tests/e2e/playwright.config.ts",
+    "test:e2e:all": "playwright test --workers=1 -c tests/e2e/playwright.config.ts",
+    "test:e2e:headed": "playwright test --headed -c tests/e2e/playwright.config.ts",
+    "test:e2e:debug": "playwright test --debug -c tests/e2e/playwright.config.ts",
+    "test:e2e:ui": "playwright test --ui -c tests/e2e/playwright.config.ts",
     "test:e2e:report": "playwright show-report",
-    "typecheck:e2e": "tsc --noEmit -p tsconfig.playwright.json",
+    "typecheck:e2e": "tsc --noEmit -p tests/e2e/tsconfig.json",
     "test:all": "npm run typecheck:e2e && npm run test:e2e:all"
   },
   "devDependencies": {
