+++ tsconfig.json
@@ -1,3 +1,3 @@
 {
-  "extends": "./tsconfig.playwright.json"
+  "extends": "./tests/e2e/tsconfig.json"
 }
